third_party/blink/web_tests/external/wpt/css/css-mixins/at-function-cssom-expected.txt

This is a testharness.js-based test.
[FAIL] Indexed property getter
  assert_array_equals: expected property 0 to be "--x" but got undefined (expected array ["--x", "result"] got [undefined, undefined])
Harness: the test ran to completion.
